Beginnings of rewrite and seperation into files
