+++ assets/s1.css
@@ -592,9 +592,10 @@ there.
     position: relative;
     box-shadow: 2px 2px 2px lightgrey;
   }
-  #plotly-image {
+  #jolyu-logo {
     margin-bottom: 0px !important;
-    height: 45px !important;
+    height: 8em !important;
+    justify-content: center;
   }
   #learn-more-button {
     margin-top: 0px !important;
