+++ helpers.py
@@ -3,6 +3,8 @@ import pandas as pd
 import math
 
 
+treetten = 21
+
 def DataToMonths(data):
     df = data.resample('M').sum()
     return(df)
